+++ src/interpupillaryDistance.js
@@ -8,8 +8,8 @@ import Arrow from './media/arrow.svg'
 import PD from './media/pd.png?width=480&height=240'
 import { bindKeys, unbindKeys } from './components/keyBinder'
 import { addButtons } from './components/buttons'
-import { colorDarkRed } from './constants'
 import text from './text.json'
+import { setDefaultVideoPosition } from './components/video'
 
 // let selfVideo = false // No WebGazer video available and an extra video element needs to be created
 
@@ -58,6 +58,8 @@ RemoteCalibrator.prototype.measurePD = function (options = {}, callback) {
     videoHeight
   )
 
+  const RC = this
+
   const breakFunction = () => {
     ruler.removeEventListener('mousedown', rulerListener)
     this._removeBackground()
@@ -71,10 +73,12 @@ RemoteCalibrator.prototype.measurePD = function (options = {}, callback) {
       height: originalStyles.videoHeight,
       width: originalStyles.videoWidth,
       opacity: originalStyles.opacity,
-      left: '10px',
-      bottom: '10px',
       borderRadius: '0px',
     })
+    setDefaultVideoPosition(
+      RC,
+      document.querySelector('#webgazerVideoContainer')
+    )
 
     Object.assign(document.querySelector('#webgazerVideoFeed').style, {
       height: originalStyles.videoHeight,
@@ -124,7 +128,7 @@ RemoteCalibrator.prototype.measurePD = function (options = {}, callback) {
   // TODO To be removed
   setTimeout(() => {
     Swal.fire({
-      ...swalInfoOptions,
+      ...swalInfoOptions(this),
       icon: undefined,
       imageUrl: PD,
       imageWidth: 480,
@@ -190,15 +194,20 @@ const formatVideo = (RC, video, canvas, container, stream = null) => {
     height: h + 'px',
     width: window.innerWidth * videoWidthFactor + 'px',
     opacity: 1,
-    // left: `50%`,
-    // top: `50%`,
-    // transform: `translate(${(-window.innerWidth * videoWidthFactor) / 2}px, ${
-    //   -h * 0.5
-    // }px)`,
-    left: 0.5 * window.innerWidth * (1 - videoWidthFactor) + 'px',
-    bottom: 0.5 * (window.innerHeight - h) + 'px',
     borderRadius: '15px',
   })
+
+  if (RC.isMobile.value) {
+    Object.assign(container.style, {
+      right: 0.5 * window.innerWidth * (1 - videoWidthFactor) + 'px',
+      top: 0.5 * (window.innerHeight - h) + 'px',
+    })
+  } else {
+    Object.assign(container.style, {
+      left: 0.5 * window.innerWidth * (1 - videoWidthFactor) + 'px',
+      bottom: 0.5 * (window.innerHeight - h) + 'px',
+    })
+  }
 
   // Canvas
   // Object.assign(canvas.style, {
@@ -277,7 +286,7 @@ const setupRuler = (RC, screenPpi, vWidth, vHeight) => {
       thisText.innerHTML = i / 10
       scales.appendChild(thisText)
 
-      if (i === 0) thisText.style.color = colorDarkRed
+      if (i === 0) thisText.style.color = RC._CONST.COLOR.DARK_RED
     }
   }
 
@@ -290,7 +299,9 @@ const setupRuler = (RC, screenPpi, vWidth, vHeight) => {
   selectionElement.style.left = '-100px'
   selectionElement.style.top = '40px'
 
-  document.getElementById('size-arrow-fill').setAttribute('fill', colorDarkRed)
+  document
+    .getElementById('size-arrow-fill')
+    .setAttribute('fill', RC._CONST.COLOR.DARK_RED)
 
   const _onDownRuler = e => {
     selectionElement.style.left = (offsetPixel = e.offsetX - 30) + 'px'